refactor: tool 패키지구조 변경
