+++ app/src/main/java/woowacourse/paint/paintBoard/painter/tool/drawable/NormalPen.kt
@@ -1,31 +1,26 @@
-package woowacourse.paint.paintBoard.tools
+package woowacourse.paint.paintBoard.painter.tool.drawable
 
 import android.graphics.Paint
-import android.graphics.PorterDuff
-import android.graphics.PorterDuffXfermode
 import woowacourse.paint.paintBoard.Brush
 import woowacourse.paint.paintBoard.Line
 
-class LineEraser(
-    private val onRemove: (Line) -> Unit,
-    private val line: Line = Line(brush = Brush(
+
+class NormalPen(
+    reset: () -> Unit,
+    private val onSave: (Line) -> Unit,
+    override val line: Line = Line(brush = Brush(
         Paint().apply {
             style = Paint.Style.STROKE
             strokeCap = Paint.Cap.ROUND
             strokeJoin = Paint.Join.ROUND
             strokeMiter = SOFT_ANGLE
             isAntiAlias = true
-            xfermode = PorterDuffXfermode(PorterDuff.Mode.CLEAR)
         }
     ))
-) : EraseTool() {
-
-    fun setWidth(width: Float) {
-        line.brush.changeBrushWidth(width)
-    }
+) : DrawableTool(line, reset) {
 
     override fun startPainting(pointX: Float, pointY: Float) {
-        onRemove(line)
+        onSave(line)
         line.path.moveTo(pointX, pointY)
         line.path.lineTo(pointX, pointY)
     }
