+++ app/src/main/java/woowacourse/paint/paintBoard/painter/tool/erase/EraseTool.kt
@@ -1,0 +1,8 @@
+package woowacourse.paint.paintBoard.painter.tool.erase
+
+import woowacourse.paint.paintBoard.painter.Painter
+
+abstract class EraseTool : Painter {
+
+    override fun finishPainting() {}
+}
