+++ app/src/main/java/woowacourse/paint/paintBoard/painter/tool/erase/LineEraser.kt
@@ -1,26 +1,31 @@
-package woowacourse.paint.paintBoard.tools
+package woowacourse.paint.paintBoard.painter.tool.erase
 
 import android.graphics.Paint
+import android.graphics.PorterDuff
+import android.graphics.PorterDuffXfermode
 import woowacourse.paint.paintBoard.Brush
 import woowacourse.paint.paintBoard.Line
 
-
-class NormalPen(
-    reset: () -> Unit,
-    private val onSave: (Line) -> Unit,
-    override val line: Line = Line(brush = Brush(
+class LineEraser(
+    private val onRemove: (Line) -> Unit,
+    private val line: Line = Line(brush = Brush(
         Paint().apply {
             style = Paint.Style.STROKE
             strokeCap = Paint.Cap.ROUND
             strokeJoin = Paint.Join.ROUND
             strokeMiter = SOFT_ANGLE
             isAntiAlias = true
+            xfermode = PorterDuffXfermode(PorterDuff.Mode.CLEAR)
         }
     ))
-) : DrawableTool(line, reset) {
+) : EraseTool() {
+
+    fun setWidth(width: Float) {
+        line.brush.changeBrushWidth(width)
+    }
 
     override fun startPainting(pointX: Float, pointY: Float) {
-        onSave(line)
+        onRemove(line)
         line.path.moveTo(pointX, pointY)
         line.path.lineTo(pointX, pointY)
     }
