--- app/src/main/java/woowacourse/paint/paintBoard/tools/EraseTool.kt
@@ -1,6 +1,0 @@
-package woowacourse.paint.paintBoard.tools
-
-abstract class EraseTool : Painter {
-
-    override fun finishPainting() {}
-}
